deploy script

+++ Kube/scripts/deploy-shield.sh
@@ -65,7 +65,10 @@ if [ "$SHIELD_FARM" == "yes" ]; then
         kubectl label node --all shield-role/farm-services=accept --overwrite
         kubectl label node --all shield-role/remote-browsers=accept --overwrite
     fi
-    curl -s -o custom-farm.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-farm.yaml
+    if [ ! -f "custom-farm.yaml" ]; then
+       curl -s -o custom-farm.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-farm.yaml
+    fi
+
     helm upgrade --install shield-farm-services shield-repo/shield --namespace=farm-services --set-string "farm-services.TZ=${TZ}" -f custom-farm.yaml --debug | tee -a "$LOGFILE"
     sleep 30
 fi
@@ -75,7 +78,9 @@ if [ "$SHIELD_MNG" == "yes" ]; then
     if [ "$SET_LABELS" == "yes" ]; then
         kubectl label node --all shield-role/management=accept --overwrite
     fi
-    curl -s -o custom-management.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-management.yaml
+    if [ ! -f "custom-management.yaml" ]; then
+       curl -s -o custom-management.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-management.yaml
+    fi
     helm upgrade --install shield-management shield-repo/shield --namespace=management --set-string "shield-mng.TZ=${TZ}" -f custom-management.yaml --debug | tee -a "$LOGFILE"
     sleep 30
 fi
@@ -85,7 +90,9 @@ if [ "$SHIELD_PROXY" == "yes" ]; then
     if [ "$SET_LABELS" == "yes" ]; then
         kubectl label node --all shield-role/proxy=accept --overwrite
     fi
-    curl -s -o custom-proxy.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-proxy.yaml
+    if [ ! -f "custom-proxy.yaml" ]; then
+        curl -s -o custom-proxy.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-proxy.yaml
+    fi
     helm upgrade --install shield-proxy shield-repo/shield --namespace=proxy --set-string "shield-proxy.TZ=${TZ}" --set-string "shield-proxy.UPSTREAM_DNS_SERVERS=${UPSTREAM_DNS_SERVERS}" -f custom-proxy.yaml --debug | tee -a "$LOGFILE"
     sleep 30
 fi
@@ -95,7 +102,9 @@ if [ "$SHIELD_ELK" == "yes" ]; then
     if [ "$SET_LABELS" == "yes" ]; then
         kubectl label node --all shield-role/elk=accept --overwrite
     fi
-    curl -s -o custom-values-elk.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-values-elk.yaml
+    if [ ! -f "custom-values-elk" ]; then
+       curl -s -o custom-values-elk.yaml https://raw.githubusercontent.com/EricomSoftwareLtd/Shield/$BRANCH/Kube/scripts/custom-values-elk.yaml
+    fi
     helm upgrade --install shield-elk shield-repo/shield --namespace=elk -f custom-values-elk.yaml --debug | tee -a "$LOGFILE"
 fi
